Fix not split picking on scheduled date set in the past
When the date_planned of the purchase order line is moved in the future
the scheduled date of the picking is not changed by the `super` call of
the `write` function.
But when the date_planned is moved in the past the scheduled date
of the picking is changed to the new date_planned.

This is happening because the `scheduled_date` on the stock.picking is
computed and the oldest `date_expected` of the related stock.moves is
set as `scheduled_date`.

https://github.com/odoo/odoo/blob/b53d36b8c4f6afd44b884f6acea9a222ff79b52d/addons/stock/models/stock_picking.py#L464

This change fixes the problem, by not accepting a picking as valid if it
has multiple date planned set on the stock.moves